Implemented A Temporal Change Of State When The Player Dies 1. When The Player Dies On The Lake, The Game Return To Main Menu (If Dies Twice And Try To Go Into The Main Menu Will Crash, Soon Will Be Fixed) 2. Corrected The Speed Of The Player During A Double Jump. 3. Implemented A Method In GameCore For Restart The Default Values Of The Camera So The Players Can Start Over In The Main Menu.

+++ ModelingProject1/SourceCode/Sprite.h
@@ -81,11 +81,10 @@ class Sprite
 
 	CollisionBox* getCollisionBox() {return spriteCollisionBox; }
 
-	int getBoxX() { return spriteCollisionBox->getX(); }
-	int getBoxY() { return spriteCollisionBox->getY(); }
-	int getBoxWidth() { return spriteCollisionBox->getWidth(); }
-	int getBoxHeight() { return spriteCollisionBox->getHeight(); }
-
+	GLfloat getBoxX() { return spriteCollisionBox->getX(); }
+	GLfloat getBoxY() { return spriteCollisionBox->getY(); }
+	GLfloat getBoxWidth() { return spriteCollisionBox->getWidth(); }
+	GLfloat getBoxHeight() { return spriteCollisionBox->getHeight(); }
 
 	bool getPlayerMoveInXCurrentFrame() { return playerMoveInXInCurrentFrame; }
 	bool getPlayerMoveInYCurrentFrame() { return playerMoveInYInCurrentFrame; }
